# Speed up create singletons