--- a/opal/_version.py
+++ b/opal/_version.py
@@ -1,4 +1,4 @@
 """
 Declare our current version string
 """
-__version__ = '0.18.2'
+__version__ = '0.18.3'