On a relatively small patient cohort 

with some singletons required to be created 
It was talking 152 sec
After this change it takes 9 seconds.

with no singletons required to be created
It was taking 76 sec
After this change it takes 0.08 seconds.